refactor(applicant): remove unnecessary code
